Move nearley_utils into lib
It doesn't actually use anything that requires it to live in the
background script, as far as I can tell.
