+++ src/lib/keyseq.ts
@@ -22,7 +22,7 @@
 
 /** */
 import { find, filter, izip } from "@src/lib/itertools"
-import { Parser } from "@src/background/nearley_utils"
+import { Parser } from "@src/lib/nearley_utils"
 import * as bracketexpr_grammar from "@src/grammars/.bracketexpr.generated"
 const bracketexpr_parser = new Parser(bracketexpr_grammar)
