+++ src/lib/nearley_utils.ts
